Use Markov and Risk Measures in ALD cuts
The code is a relatively simple copy of the original modifyvaluefunction!
from SDDP.jl, but we still don't write the cuts, since they need further
info (\rho). that is not compatible with the current cut format.

Because of this separation, add_aldcut!() depends on more state: both
xhat and niter are now taken from the SDDPModel "m" passed as argument
(which was previously unused).

This commit also reintroduces @timeit for parts of the algorithm, which
changes identation, so using -w can help in parts.