New configuration for ALD
Now most parameters for ALD are in ALDparams structure, which can be
changed as the algorithm evolves (but probably no on every iteration).
They contain the Lipschitz constant for the currnt subproblem, and flags
to control usage of lipschitz tents or ALD cuts, and the size of the ALD
cut buffer.

This comes with a more rigid "rho policy", that enforces a linear increase,
but is easier to configure by the user.

Side effects:
- The solver logic also was slightly altered:
  * multipliers are either zero (for Lip tents) or obtained by solving
    the LP relaxation
  * to accomodate new parameters controlling rho, its value is decided
    further down: now at the subproblem level
- Split functions prepareALD! and setALDsolver! from solver.jl into setup.jl
- make_cut also migrated to utils.jl, since it is not a part of solve